+++ example/integration_test/camera_update_test.dart
@@ -1,0 +1,33 @@
+import 'dart:async';
+
+import 'package:flutter/material.dart';
+import 'package:flutter_naver_map/flutter_naver_map.dart';
+import 'package:flutter_test/flutter_test.dart';
+
+import 'util/test_util.dart';
+
+void main() {
+  testNaverMap("Camera Update With Reason Test", (controller, tester) async {
+    (NCameraUpdateReason, bool)? lastMovedData;
+    final onCameraChangeStreamSubscription = tester
+        .testPageState.onCameraChangeStreamController.stream
+        .listen((event) {
+      lastMovedData = event;
+      print("[moved] $event");
+    });
+
+    final cameraUpdate = NCameraUpdate.zoomIn()
+      ..setAnimation(duration: Duration.zero)
+      // setReason Method visibility is only testing.
+      ..setReason(NCameraUpdateReason.control); // 위치 UI는 location.
+
+    await controller.updateCamera(cameraUpdate);
+    await Future.delayed(const Duration(milliseconds: 100));
+
+    print(lastMovedData);
+
+    expect(lastMovedData?.$1, NCameraUpdateReason.control);
+    expect(lastMovedData?.$2, false);
+    onCameraChangeStreamSubscription.cancel();
+  });
+}