+++ example/lib/main_test.dart
@@ -45,19 +45,30 @@ class TestPage extends StatefulWidget {
 @visibleForTesting
 class TestPageState extends State<TestPage> {
   final Completer<NaverMapController> mapControllerCompleter = Completer();
+  final onCameraChangeStreamController =
+      StreamController<(NCameraUpdateReason, bool)>.broadcast();
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
         body: NaverMap(
-            options: const NaverMapViewOptions(
-                indoorEnable: true,
-                locationButtonEnable: true,
-                consumeSymbolTapEvents: false),
-            onMapReady: (controller) async {
-              mapControllerCompleter.complete(controller);
-              log("onMapReady", name: "onMapReady");
-            }));
+      options: const NaverMapViewOptions(
+          initialCameraPosition:
+              NCameraPosition(target: NLatLng(37, 127), zoom: 14),
+          indoorEnable: true,
+          locationButtonEnable: true,
+          consumeSymbolTapEvents: false),
+      onMapReady: (controller) async {
+        mapControllerCompleter.complete(controller);
+        log("onMapReady", name: "onMapReady");
+      },
+      onCameraIdle: () {
+        print("onCameraIdle: ${DateTime.now().millisecondsSinceEpoch}");
+      },
+      onCameraChange: (reason, animated) {
+        onCameraChangeStreamController.add((reason, animated));
+      },
+    ));
   }
 
   void newMapTestPage(String tag) {