+++ example/lib/util/overlay_portal_util.dart
@@ -1,12 +1,15 @@
+import 'package:balloon_widget/balloon_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_naver_map/flutter_naver_map.dart';
-import 'package:flutter_naver_map_example/design/custom_widget.dart';
+
+// import 'package:flutter_naver_map_example/design/custom_widget.dart';
 import 'package:flutter_naver_map_example/design/theme.dart';
 
 typedef NInfoOverlayBuilderCreateCallback = Widget Function(
     BuildContext context,
     NaverMapController mapController,
-    NInfoOverlayPortalController controller);
+    NInfoOverlayPortalController controller,
+    VoidCallback back);
 
 typedef NInfoOverlayBuilderCallback = Widget Function(
     BuildContext context, NaverMapController mapController);
@@ -22,33 +25,49 @@ class NInfoOverlayPortalController extends OverlayPortalController {
   void openWithWidget({
     required NInfoOverlayBuilderCreateCallback builder,
     required NPoint screenPoint,
+    required Stream<NPoint> screenPointStream,
     required NOverlay<dynamic> overlay,
   }) {
     _builder = (BuildContext context, NaverMapController mapController) =>
         _nOverlayInfoOverlayWithTouchScope(context,
             point: screenPoint,
-            child: builder.call(context, mapController, this));
+            screenPointStream: screenPointStream,
+            child: builder.call(context, mapController, this, hide));
     show();
   }
 
   Widget _nOverlayInfoOverlayWithTouchScope(BuildContext context,
-      {required NPoint point, required Widget child}) {
-    final xPosition = (point.x) - 28;
+      {required NPoint point,
+      required Stream<NPoint> screenPointStream,
+      required Widget child}) {
     return Positioned.fill(
-        child: GestureDetector(
-            behavior: HitTestBehavior.opaque,
-            onTap: () => hide(),
-            child: Container(
-                color: Colors.black12,
-                child: Stack(children: [
-                  Positioned(
-                      left: xPosition < 0 ? 0 : xPosition,
-                      top: point.y < 0 ? 0 : point.y,
-                      child: Balloon(
-                          size: const Size(200, 160),
-                          padding: EdgeInsets.zero,
-                          backgroundColor: getColorTheme(context).background,
-                          child: child)),
-                ]))));
+        child: StreamBuilder<NPoint>(
+            stream: screenPointStream,
+            builder: (context, snapshot) {
+              final xPosition = snapshot.data?.x ?? point.x;
+              final yPosition = snapshot.data?.y ?? point.y;
+              final screenSize = MediaQuery.sizeOf(context);
+              if ((xPosition < 0 || screenSize.width < xPosition) ||
+                  (yPosition < 0 || screenSize.height < yPosition)) {
+                WidgetsBinding.instance.addPostFrameCallback((_) => hide());
+              }
+              return Stack(children: [
+                Positioned(
+                    left: xPosition - 28,
+                    top: yPosition,
+                    child: Balloon(
+                        borderRadius: BorderRadius.circular(12),
+                        nipMargin: 8,
+                        nipSize: 16,
+                        nipRadius: 4,
+                        elevation: 6,
+                        padding: EdgeInsets.zero,
+                        color: getColorTheme(context).background,
+                        shadowColor: Colors.black,
+                        nipPosition: BalloonNipPosition.topLeft,
+                        child:
+                            SizedBox(width: 200, height: 160, child: child))),
+              ]);
+            }));
   }
 }