+++ lib/flutter_naver_map.dart
@@ -16,6 +16,7 @@ import 'package:flutter_naver_map/src/messaging/messaging.dart';
 import 'package:flutter_naver_map/src/util/image_util.dart';
 import 'package:flutter_naver_map/src/util/math.dart';
 import 'package:flutter_naver_map/src/util/widget_to_image.dart';
+import "package:meta/meta.dart";
 
 /*
   --- controller ---
@@ -124,3 +125,5 @@ part 'src/type/map/overlay/overlay_image.dart';
 part 'src/widget/map_widget.dart';
 
 part 'src/widget/platform_view.dart';
+
+part 'src/widget/control_widget/zoom_control_widget.dart';