+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## 1.2.0
+
+### Improve
+- [All Platform] Add / Improve APIs
+  - Add property `NaverMapController.nowCameraPosition` as a experimental api which type is `NCameraPosition` (non-async)
+  - Add method `NCameraUpdate.setReason`
+  - Change return type of method `NaverMapController.getMeterPerDp` 
+      & `NaverMapController.getMeterPerDpAtLatitude` to non-async (Future<double> -> double)
+  - Change `NCameraUpdate.setAnimation` to have non-null parameters with default values
+  - Change property type `NaverMapViewOptions.initialCameraPosition` to non-nullable type
+
+### Breaking Change
+- [Flutter] Change support minimum Flutter SDK Version to 3.19.0, Dart 3.0.0 
+
+### Fix
+- [Android] Change Platform View display mode to TLHC (related issue [#152](https://github.com/note11g/flutter_naver_map/issues/152))
+- [Android] Change Flutter Render View to SurfaceView at Android 11~13 (related issue [#152](https://github.com/note11g/flutter_naver_map/issues/152))
+
 ## 1.1.2
 
 ### Improve